# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aurangabad Tours & Travel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tour {
  id: number;
  name: string;
  /** cave, fort, temple, garden, lake, hill, mosque, mausoleum */
  category: string;
  description: string;
  location: string;
  duration: string;
  highlights: string[];
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  price?: number | null;
}

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 10 */
  phone: string;
  /**
   * ID of the specific tour, null if general inquiry
   * @nullable
   */
  tourId: number | null;
  /** @nullable */
  tourName?: string | null;
  /**
   * Mumbai, Pune, or Aurangabad
   * @nullable
   */
  pickupLocation?: string | null;
  /** @nullable */
  travelDate?: string | null;
  /** @nullable */
  groupSize?: number | null;
  /** @minLength 1 */
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  tourId?: number | null;
  /** @nullable */
  tourName?: string | null;
  /** @nullable */
  pickupLocation?: string | null;
  /** @nullable */
  travelDate?: string | null;
  /** @nullable */
  groupSize?: number | null;
  message: string;
  createdAt: string;
}

export interface TransportOption {
  id: number;
  from: string;
  to: string;
  vehicleType: string;
  capacity: number;
  pricePerPerson: number;
  duration: string;
  /** @nullable */
  description?: string | null;
}

export interface ErrorResponse {
  error: string;
}