# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aurangabad Tours & Travel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours/{id}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get a specific tour by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "404":
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Tour:
      type: object
      required: [id, name, category, description, location, duration, highlights, imageUrl, featured]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "cave, fort, temple, garden, lake, hill, mosque, mausoleum"
        description:
          type: string
        location:
          type: string
        duration:
          type: string
        highlights:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        rating:
          type: ["number", "null"]
        price:
          type: ["number", "null"]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string