# lib/api-zod/src/generated/types/tour.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aurangabad Tours & Travel API
 * OpenAPI spec version: 0.1.0
 */

export interface Tour {
  id: number;
  name: string;
  /** cave, fort, temple, garden, lake, hill, mosque, mausoleum */
  category: string;
  description: string;
  location: string;
  duration: string;
  highlights: string[];
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  price?: number | null;
}